improve variables managment

+++ glaber.sh
@@ -97,7 +97,6 @@ usage() {
   echo "$0 stop     - Stop glaber containers"
   echo "$0 recreate - Completely remove glaber and start it again"
   echo "$0 remove   - Completely remove glaber installation"
-  echo "$0 remote   - Remote rebuild github glaber images (only admins)"
   echo "$0 diag     - Collect glaber start and some base system info to the file"
 }
 
@@ -133,10 +132,14 @@ stop() {
 }
 
 remove() {
-  docker-compose down
-  # are you shure?
-  rm .passwords.created .zbxweb || true
-  sudo rm -rf  mysql/mysql_data/ clickhouse/clickhouse_data
+  read -p "Are you sure than you want completely remove glaber with database data [y/n] ? " -n 1 -r
+  echo
+  if [[ $REPLY =~ ^[Yy]$ ]]
+  then
+    docker-compose down
+    rm .passwords.created .zbxweb || true
+    sudo rm -rf  mysql/mysql_data/ clickhouse/clickhouse_data
+  fi
 }
 
 recreate() {
